Re-organise the codes
-Put all ``struct`` definitions into header files
-Delete **tools.h**
