+++ sector_cext.h
@@ -1,3 +1,62 @@
+#include<stdio.h>
+#include<stdlib.h>
+#include<string.h>
+#include<math.h>
+#include<time.h>
+
+/* * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
+ *                                                                             *
+ * Basic functions                                                             *
+ *                                                                             *
+ * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
+FILE *open_file(char *fName, char *mode);
+
+double **malloc_2d_double(int nRow, int nCol);
+double **memcpy_2d_double(double **source, int nRow, int nCol);
+void free_2d_double(double **p, int nRow);
+
+int bisection_search(double a, double *x, int nX);
+double interp(double xp, double *x, double *y, int nPts);
+double trapz_table(double *y, double *x, int nPts, double a, double b);
+double trapz_filter(double *filter, double *flux, double *waves, int nWaves);
+
+struct linResult {
+    double slope;
+    double intercept;
+    double R;
+};
+struct linResult linregress(double *x, double *y, int nPts);
+
+
+/* * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
+ *                                                                             *
+ * SFHs related                                                                *
+ *                                                                             *
+ * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
+struct ssp {
+    short index;
+    float metals;
+    float sfr;
+};
+struct csp {
+    struct ssp *bursts;
+    int nBurst;
+};
+struct gal_params {
+    double z;
+    int nAgeStep;
+    double *ageStep;
+    int nGal;
+    int *indices;
+    struct csp *histories;
+};
+
+
+/* * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
+ *                                                                             *
+ * SEDs and dust related                                                       *
+ *                                                                             *
+ * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
 struct sed_params {
     // Raw templates
     int minZ;
@@ -30,17 +89,6 @@ struct sed_params {
     double *outBC;
 };
 
-void init_templates_raw(struct sed_params *spectra, char *fName);
-void shrink_templates_raw(struct sed_params *spectra, double maxAge);
-
-
-//void init_filters(struct sed_params *spectra, double *filters, int nFlux, int nObs,
-//                  double *logWaves, double *LyAbsorption);
-
-
-//void free_filters(struct sed_params *spectra);
-
-
 struct dust_params {
     double tauUV_ISM;
     double nISM;
@@ -49,30 +97,10 @@ struct dust_params {
     double tBC;
 };
 
-
-struct ssp {
-    short index;
-    float metals;
-    float sfr;
-};
-
-
-struct csp {
-    struct ssp *bursts;
-    int nBurst;
-};
-
-
-struct gal_params {
-    double z;
-    int nAgeStep;
-    double *ageStep;
-    int nGal;
-    int *indices;
-    struct csp *histories;
-};
-
-
+#ifndef _SECTOR_
+void init_templates_raw(struct sed_params *spectra, char *fName);
+void shrink_templates_raw(struct sed_params *spectra, double maxAge);
 double *composite_spectra_cext(struct sed_params *spectra,
                                struct gal_params *galParams, struct dust_params *dustParams,
                                short outType, short nThread);
+#endif
